Everything is a CheckedAtoB
